feat(input): added input

+++ include/io/video.hpp
@@ -1,14 +1,12 @@
 #pragma once
 
-#include <stdint.h>
-
 namespace chip::IO {
 
 class Video {
 public:
     static void clear(bool value);
-    static void setPixel(uint8_t x, uint8_t y, bool value);
-    static bool getPixel(uint8_t x, uint8_t y);
+    static void setPixel(int x, int y, bool value);
+    static bool getPixel(int x, int y);
 private:
     Video() {}
     static bool framebuffer_[32][64];
